requirements.txt
python-telegram-bot==20.7
yt-dlp


💡 Ide Proyek: Telegram Video Downloader Bot
🎯 Fungsi Bot
User kirim URL video (TikTok, Instagram Reels, YouTube Shorts, dll)

Bot akan:

Mendapatkan video dari URL tersebut

Mengunduh video (tanpa watermark kalau bisa)

Mengirim video tersebut kembali ke Telegram

🔧 Cara Implementasi Berdasarkan Platform
✅ 1. TikTok
Tanpa API Resmi — Kamu bisa gunakan:

TikTokAPI-Py (butuh cookies)

Atau gunakan layanan pihak ketiga seperti:

https://ttdownloader.com/

https://snaptik.app/

Metode:

Kirim request ke URL layanan (dengan scraping/form request)

Dapatkan link download dari response HTML

Ambil link tanpa watermark

Unduh file .mp4

Kirim ke Telegram bot

✅ 2. Instagram Reels
Sulit pakai scraping langsung karena perlu login dan CSRF token

Alternatif:

Gunakan instaloader untuk private account (CLI, bisa disesuaikan ke Python)

Scrape dari situs seperti igram.io atau instadownloader.org

✅ 3. YouTube (Shorts / Video Umum)
Gunakan yt-dlp (rekomendasi utama)

Contoh:

python
Copy
Edit
import yt_dlp

def download_youtube(url, filename="video.mp4"):
    ydl_opts = {
        'outtmpl': filename,
        'format': 'best[ext=mp4]',
    }
    with yt_dlp.YoutubeDL(ydl_opts) as ydl:
        ydl.download([url])
📦 Struktur Bot Telegram Sederhana
bash
Copy
Edit
video_bot/
├── bot.py
├── downloader/
│   ├── tiktok.py
│   ├── instagram.py
│   └── youtube.py
├── requirements.txt
└── README.md
bot.py: Handler Telegram + parsing URL + call downloader

downloader/*.py: Fungsi berdasarkan domain

requirements.txt: python-telegram-bot, yt-dlp, requests, dll

🧠 Deteksi Platform Otomatis
Kamu bisa gunakan regex untuk cek domain URL:

python
Copy
Edit
if "tiktok.com" in url:
    video_path = download_tiktok(url)
elif "youtube.com" in url or "youtu.be" in url:
    video_path = download_youtube(url)
elif "instagram.com" in url:
    video_path = download_instagram(url)
else:
    bot.send_message(chat_id, "URL tidak dikenali.")
🛡️ Catatan Penting
Beberapa layanan scraping pihak ketiga memiliki rate limit atau captcha

Pastikan ada pengecekan error: URL tidak valid, video private, dll

Pastikan video size tidak lebih dari 50MB (Telegram limit)

Gunakan async atau worker queue jika ingin support multiple download